Skyward log file export

+++ custom/src/YExportFiles.cc
@@ -13,6 +13,7 @@
 //-----------------------------------------------------------------------------
 YExportFiles::YExportFiles()
     : _convertToUTM(false)
+    , _convertToSkyward(false)
     , _cancel(false)
     , _totalFiles(0)
     , _curFile(0)
@@ -27,9 +28,10 @@ YExportFiles::~YExportFiles()
 
 //-----------------------------------------------------------------------------
 bool
-YExportFiles::exportData(bool convertToUTM)
+YExportFiles::exportData(bool convertToUTM, bool convertToSkyward)
 {
     _convertToUTM = convertToUTM;
+    _convertToSkyward = convertToSkyward;
     _totalFiles = 0;
     _curFile = 0;
     this->start(QThread::HighPriority);
@@ -61,7 +63,9 @@ YExportFiles::run()
         return;
     }
     _totalFiles = _filesInPath(telemetryPath);
-    if(_convertToUTM) {
+    if(_convertToUTM && _convertToSkyward) {
+        _totalFiles *= 3;
+    } else if(_convertToUTM || _convertToSkyward) {
         _totalFiles *= 2;
     }
     _totalFiles += _filesInPath(missionPath);
@@ -75,7 +79,7 @@ YExportFiles::run()
         if(!_cancel && _copyFilesInPath(telemetryPath, QString("%1/%2").arg(targetPath).arg(AppSettings::telemetryDirectory))) {
             //-- Save UTM files
             ok = true;
-            if(!_cancel && _convertToUTM) {
+            if(!_cancel && (_convertToUTM || _convertToSkyward)) {
                 emit message(QString(tr("Exporting UTM telemetry files...")));
                 if(!_convertLogsToUTM(telemetryPath, QString("%1/%2").arg(targetPath).arg(AppSettings::telemetryDirectory))) {
                     ok = false;
@@ -143,18 +147,34 @@ YExportFiles::_convertLogsToUTM(const QString src, const QString dst)
     QDirIterator it(src, QStringList() << "*.tlog", QDir::Files, QDirIterator::NoIteratorFlags);
     while(it.hasNext() && !_cancel) {
         QFileInfo fi(it.next());
-        QString output = dst + "/" + fi.baseName() + QStringLiteral(".utm");
-        QFileInfo fo(output);
-        if(fo.exists()) {
-            QFile::remove(fo.filePath());
+        if(!_cancel && _convertToUTM) {
+            QString output = dst + "/" + fi.baseName() + QStringLiteral(".utm");
+            QFileInfo fo(output);
+            if(fo.exists()) {
+                QFile::remove(fo.filePath());
+            }
+            UTMConverter converter(false);
+            connect(this, &YExportFiles::cancelProcess, &converter, &UTMConverter::cancel);
+            if(!converter.convertTelemetryFile(fi.filePath(), fo.filePath())) {
+                emit message(QString(tr("Error exporting %1")).arg(fo.filePath()));
+                return false;
+            }
+            emit copyCompleted(_totalFiles, ++_curFile);
         }
-        UTMConverter converter;
-        connect(this, &YExportFiles::cancelProcess, &converter, &UTMConverter::cancel);
-        if(!converter.convertTelemetryFile(fi.filePath(), fo.filePath())) {
-            emit message(QString(tr("Error exporting %1")).arg(fo.filePath()));
-            return false;
+        if(!_cancel && _convertToSkyward) {
+            QString output = dst + "/" + fi.baseName() + QStringLiteral(".sky");
+            QFileInfo fo(output);
+            if(fo.exists()) {
+                QFile::remove(fo.filePath());
+            }
+            UTMConverter converter(true);
+            connect(this, &YExportFiles::cancelProcess, &converter, &UTMConverter::cancel);
+            if(!converter.convertTelemetryFile(fi.filePath(), fo.filePath())) {
+                emit message(QString(tr("Error exporting %1")).arg(fo.filePath()));
+                return false;
+            }
+            emit copyCompleted(_totalFiles, ++_curFile);
         }
-        emit copyCompleted(_totalFiles, ++_curFile);
     }
     return true;
 }
